fix: channel naming is overriden
