+++ Dockerfile.response
@@ -5,9 +5,9 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && rm -rf /var/lib/apt/lists/*
 
 WORKDIR /app
-COPY requirements.txt /app
-RUN pip install -r requirements.txt
 
 COPY . /app/
 
-CMD ["bash", "startup.sh"]
+RUN pip install -r requirements.txt
+
+CMD ["bash", "/app/startup.sh"]
