+++ incidentresponse/actions/incident_commands.py
@@ -1,6 +1,5 @@
 from response.core.models import Incident
-from response.slack.decorators.incident_command import incident_command
-from response.slack.decorators import get_help
+from response.slack.decorators.incident_command import incident_command, get_help
 from response.slack.client import reference_to_id, SlackError
 from response.slack.client import SlackClient
 from datetime import datetime
